+++ src/SafeRegex/Internal/Errors/Errors/CompileError.php
@@ -1,8 +1,0 @@
-<?php
-namespace TRegx\SafeRegex\Internal\Errors\Errors;
-
-use TRegx\SafeRegex\Internal\Errors\HostError;
-
-interface CompileError extends HostError
-{
-}