+++ test/Utils/Runtime/CausesWarnings.php
@@ -12,11 +12,4 @@ trait CausesWarnings
     {
         @\preg_match('/unclosed pattern', '');
     }
-
-    public function causeCompileWarning(): void
-    {
-        \preg_replace_callback('/a/', function () {
-            @\trigger_error('preg_match() error', E_USER_NOTICE);
-        }, 'a');
-    }
 }